séparation en deux fichiers des fonctions de hash password et de compare password
