+++ back/src/utils/comparePassword.ts
@@ -1,0 +1,8 @@
+import bcrypt from "bcrypt";
+
+export function comparePassword(
+  password: string,
+  hashedPassword: string
+): Promise<boolean> {
+  return bcrypt.compare(password, hashedPassword);
+}
