+++ back/src/utils/hashPassword.ts
@@ -5,10 +5,3 @@ export async function hashPassword(password: string): Promise<string> {
 
   return bcrypt.hash(password, salt);
 }
-
-export function comparePassword(
-  password: string,
-  hashedPassword: string
-): Promise<boolean> {
-  return bcrypt.compare(password, hashedPassword);
-}
